Remove changes to fdbadminclient interface to not introduce breaking changes
