+++ controllers/replace_failed_process_groups.go
@@ -22,6 +22,7 @@ package controllers
 
 import (
 	"context"
+
 	fdbv1beta2 "github.com/FoundationDB/fdb-kubernetes-operator/api/v1beta2"
 	"github.com/FoundationDB/fdb-kubernetes-operator/internal"
 	"github.com/FoundationDB/fdb-kubernetes-operator/internal/replacements"
