+++ controllers/update_database_configuration.go
@@ -23,6 +23,7 @@ package controllers
 import (
 	"context"
 	"fmt"
+
 	"k8s.io/utils/pointer"
 
 	fdbtypes "github.com/FoundationDB/fdb-kubernetes-operator/api/v1beta2"
