+++ internal/replacements/replacements.go
@@ -23,6 +23,7 @@ package replacements
 import (
 	"context"
 	"fmt"
+
 	"github.com/FoundationDB/fdb-kubernetes-operator/pkg/podmanager"
 	"sigs.k8s.io/controller-runtime/pkg/client"
