+++ internal/replacements/replacements_test.go
@@ -23,6 +23,7 @@ package replacements
 import (
 	"context"
 	"fmt"
+
 	"github.com/FoundationDB/fdb-kubernetes-operator/pkg/podmanager"
 	ctrlClient "sigs.k8s.io/controller-runtime/pkg/client"
